requirements.txt
Flask
Pillow
inky
